feat: 接口升级，与以前版本不兼容。增加tibco监听和cm功能

+++ tibco/envioment.go
@@ -10,7 +10,7 @@ import (
 
 func tibrvOpen() error {
 	if status := C.tibrv_Open(); status != C.TIBRV_OK {
-		return fmt.Errorf("Tibco Open Error, code: %d", C.uchar(status))
+		return fmt.Errorf("Tibco Open Error, code: %d", status)
 	}
 
 	return nil
@@ -18,7 +18,7 @@ func tibrvOpen() error {
 
 func tibrvClose() error {
 	if status := C.tibrv_Close(); status != C.TIBRV_OK {
-		return fmt.Errorf("Tibco Close Error, code: %d", C.uchar(status))
+		return fmt.Errorf("Tibco Close Error, code: %d", status)
 	}
 
 	return nil
